Fix integer division before float cast in SincFilter::getBufferSize
m_coeffs.size() / m_increment divided as size_t and only then cast to
float, truncating the ratio before it scaled maxRatio. No real kernel
table divides evenly by its increment, so the delay-line buffer sized
for SrPushConverter/SrPullConverter came out undersized whenever the
result cleared the 4096-sample floor (e.g. sinc_69_768.h at
SrConvertMaxRatio=128 comes out 120 samples short across 2 channels).
Cast each operand to float before dividing.

Add SincFilter_test.cpp: one case reproducing the sinc_69_768.h/128x
scenario, checked against both the correct and the old truncated
formula (verified it fails on the old formula), and one case
confirming the 4096 floor still applies when the ratio stays under it.

+++ src/includes/Filters/Sinc/SincFilter.h
@@ -66,8 +66,8 @@ class SincFilter
     /// Heuristic: three kernel widths scaled by the ratio, with a 4096-sample floor.
     [[nodiscard]] size_t getBufferSize(const float maxRatio, const size_t channels) const noexcept
     {
-        const auto width =
-            static_cast<size_t>(3 * lrint(static_cast<float>(m_coeffs.size() / m_increment) * maxRatio) + 1);
+        const auto width = static_cast<size_t>(
+            3 * lrint(static_cast<float>(m_coeffs.size()) / static_cast<float>(m_increment) * maxRatio) + 1);
         return 1 + channels * (1 + std::max<size_t>(width, 4096));
     }
